pyproject.toml
[project]
name = "loommind"
version = "0.1.0"
description = "Minimal LangGraph agent demo"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=0.2.55",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
    "lark-oapi>=1.5.3",
    "python-socks[asyncio]>=2.4.0",
]

[dependency-groups]
dev = ["ruff>=0.9.0"]

[tool.ruff]
target-version = "py313"
line-length = 88
src = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["api", "graph_agent", "feishu_app", "parser"]
